+++ testkit/src/main/java/io/takari/bpm/testkit/DeploymentProcessor.java
@@ -1,0 +1,10 @@
+package io.takari.bpm.testkit;
+
+import io.takari.bpm.xml.ParserException;
+
+import java.io.InputStream;
+
+public interface DeploymentProcessor {
+
+    void handle(InputStream in, TestProcessDefinitionProvider provider) throws ParserException;
+}