+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## [Unreleased]
 
+### Added
+- more fine-grained control over process deployment in `EngineRule`.
+- minor logging improvements.
+
 ### Changed
 - clarify the javadoc on the `ProcessDefinitionProvider#getById` method: it returns `null` if process not found instead
 of throwing an exception.
+
+
 
 ## [0.8.4] - 2016-08-29